requirements.txt
BeautifulSoup4
requests
alpaca-trade-api
pandas
yfinance